Moving all definitions to global.h + adding comments
